+++ app/src/main/java/com/androsaver/source/NextcloudSource.kt
@@ -2,7 +2,6 @@ package com.androsaver.source
 
 import android.content.Context
 import android.util.Log
-import androidx.preference.PreferenceManager
 import com.androsaver.BuildConfig
 import com.androsaver.awaitResponse
 import com.androsaver.HttpClients